+++ metro-extensions/contribute/impl-compiler-plugin/src/test/resources/dump/contributesrobot/constructorParametersWithoutInject.kt
@@ -1,0 +1,15 @@
+// RUN_PIPELINE_TILL: BACKEND
+package com.test
+
+import software.amazon.app.platform.inject.robot.ContributesRobot
+import software.amazon.app.platform.robot.Robot
+
+class RobotDependency
+
+class AnotherRobotDependency
+
+@ContributesRobot(AppScope::class)
+class TestRobot(
+  val dependency: RobotDependency,
+  val anotherDependency: AnotherRobotDependency,
+) : Robot